+++ valhalla/thor/timedistancematrix.h
@@ -55,6 +55,9 @@ public:
                  const float max_matrix_distance,
                  const uint32_t matrix_locations = kAllLocations,
                  const bool invariant = false) {
+
+    LOG_INFO("matrix::TimeDistanceMatrix");
+
     // Set the mode and costing
     mode_ = mode;
     costing_ = mode_costing[static_cast<uint32_t>(mode_)];
@@ -229,6 +232,7 @@ protected:
                           const baldr::DirectedEdge* edge,
                           const graph_tile_ptr& tile,
                           const sif::EdgeLabel& pred,
+                          const baldr::TimeInfo& time_info,
                           const uint32_t matrix_locations);
 
   /**